requirements.txt
amqp==5.0.7
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
black==21.12b0
brew==0.1.3
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.28.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0
django-celery-results==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.1
kombu==5.2.2
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi==1.1.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==1.5
pytz==2021.3
redis==4.0.2
requests==2.26.0
routers==0.10.1
ruamel.yaml==0.17.17
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
